test: return transactions
