+++ .github/workflows/integration-tests-api.yml
@@ -22,8 +22,7 @@ jobs:
           - logs.test.ts
           - stats.test.ts
           - tokens.test.ts
-          - transactions_set1.test.ts
-          - transactions_set2.test.ts
+          - transactions.test.ts
     name: 'API test set: ${{ matrix.test-pattern}} / Node: ${{ matrix.node-version}}'
     steps:
       - name: Checkout with Submodule
